consolidate request config
